Add class survey-group-table
This is the simplest patch which implements survey-groups in a compact table form.

The method render-group's api was changed so that different implementations could be selected based on the type of survey-group (instead of the survey-ctrl object).

Used survey-group-table in sf36.

darcs-hash:20100111181806-dd2a1-c311ac39fecec9be852d5546a38144186f8fcbe3.gz